@@ -23,10 +23,10 @@
   },
   "homepage": "https://github.com/vasinkevych/react-native-node#readme",
   "dependencies": {
-    "axios": "^0.15.3",
+    "axios": "^0.18.0",
     "babel-core": "^6.17.0",
     "babel-loader": "^6.2.0",
-    "babel-plugin-add-module-exports": "^0.1.2",
+    "babel-plugin-add-module-exports": "1.0.0",
     "babel-plugin-react-html-attrs": "^2.0.0",
     "babel-plugin-transform-class-properties": "^6.3.13",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
@@ -35,28 +35,29 @@
     "babel-preset-stage-0": "^6.3.13",
     "bootstrap-css-only": "^4.1.1",
     "classnames": "^2.2.5",
-    "dotenv": "^6.2.0",
+    "dotenv": "7.0.0",
     "koa": "^2.0.0-alpha.8",
     "koa-static": "^5.0.0",
     "koa2-router": "^1.0.6",
     "nodemailer": "^5.1.1",
-    "pm2": "^3.2.9",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
-    "react-redux": "^5.0.2",
-    "react-router": "^3.0.2",
-    "redux": "^3.6.0",
-    "redux-logger": "^2.8.1",
-    "redux-promise-middleware": "^4.2.0",
-    "redux-saga": "^0.14.3",
+    "pm2": "3.4.0",
+    "prop-types": "^15.7.2",
+    "react": "16.8.4",
+    "react-dom": "16.8.4",
+    "react-redux": "6.0.1",
+    "react-router-dom": "^4.4.0",
+    "redux": "4.0.1",
+    "redux-logger": "3.0.6",
+    "redux-promise-middleware": "6.1.0",
+    "redux-saga": "1.0.2",
     "redux-thunk": "^2.2.0",
-    "reselect": "^2.5.4",
+    "reselect": "4.0.0",
     "webpack": "^1.12.9",
     "webpack-dev-server": "^1.14.1"
   },
   "devDependencies": {
     "chai": "^4.2.0",
-    "eslint": "^5.13.0",
+    "eslint": "5.15.2",
     "eslint-config-koa": "^2.0.2",
     "eslint-config-standard": "^12.0.0",
     "eslint-plugin-import": "^2.16.0",
@@ -64,14 +65,14 @@
     "eslint-plugin-node": "^8.0.1",
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-standard": "^4.0.0",
-    "mocha": "^5.2.0",
+    "mocha": "6.0.2",
     "nodemon": "^1.18.10",
     "babel-core": "^6.23.1",
     "babel-loader": "^6.3.2",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-stage-0": "^6.22.0",
-    "css-loader": "^1.0.1",
+    "css-loader": "2.1.1",
     "style-loader": "^0.23.1"
   }
 }
