@@ -1,5 +1,6 @@
 import React from "react";
 import {connect} from "react-redux";
+import { withRouter } from 'react-router-dom';
 
 class Users extends React.Component {
   render() {
@@ -7,7 +8,7 @@ class Users extends React.Component {
   }
 }
 
-export default connect(
+export default withRouter(connect(
   (state) => ({}),
   {}
-)(Users);
+)(Users));
